requirements.txt
numpy==1.18.1
matplotlib==3.3.1
trajectory_planning_helpers==0.76
casadi==3.5.1
tikzplotlib==0.9.2
pyzmq==18.1.1
ad_interface_functions==0.19
scipy==1.3.3
scikit-learn==0.23.1
